+++ reusables/browser.py
@@ -1,13 +1,46 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
+#
+# Part of the Reusables package.
+#
+# Copyright (c) 2014-2016 - Chris Griffith - MIT License
+"""
+Classes for managing Firefox and Chrome Cookies
+"""
 
-import os
-import sqlite3
-import datetime
-import sys
+import os as _os
+import sqlite3 as _sqlite3
+import datetime as _dt
+import sys as _sys
+
+from .log import get_logger
+
+logger = get_logger(__name__)
+
+
+def _to_secs(td):
+    """Compatibility for 2.6 and 3.2 for timedelta.total_seconds()"""
+    if (_sys.version_info[0:2] < (2, 7) or
+       (2, 7) < _sys.version_info[0:2] < (3, 3)):
+        return ((td.microseconds + (td.seconds + td.days * 24 * 3600) * 10 **
+                6) / 10 ** 6)
+    else:
+        return td.total_seconds()
+
+
+class BrowserException(Exception):
+    """Generic parent exception for errors in this module"""
+
+
+class InvalidSchema(BrowserException):
+    """The provided DB did not have a valid scheme"""
 
 
 class CookieManager(object):
+    """
+    Parent class for all cookies managers for better cross browser
+    compatibilities and portability.
+    """
     _valid_structure = {"tables": [],
                         "columns": []}
     _db_paths = {
@@ -17,12 +50,18 @@ class CookieManager(object):
     _insert = ""
     db = ""
     table_name = ""
+    supported_versions = ()
 
     def __init__(self, db=None):
         self.db = db if db else self.find_db()
         self.verify_schema()
 
-    def _correct_tables_and_rows(self, cur):
+    def _correct_tables_and_titles(self, cur):
+        """Performs table and column name lookup and makes sure they match the
+        defined valid schema.
+        :param cur: SQLite cursor
+        :type cur: sqlite3.Connection.cursor"""
+
         grab_ts = cur.execute("SELECT name FROM sqlite_master "
                               "WHERE type = 'table'")
         tables = grab_ts.fetchall()
@@ -31,60 +70,71 @@ class CookieManager(object):
         cols = [description[0] for description in grab_row.description]
 
         if tables != self._valid_structure["tables"]:
-            raise AssertionError("Expected {} - Got {}".format(
+            raise InvalidSchema("Tables expected {} - Got {}".format(
                 self._valid_structure["tables"], tables))
         if cols != self._valid_structure["columns"]:
-            raise AssertionError("Expected {} - Got {}".format(
+            raise InvalidSchema("Columns expected {} - Got {}".format(
                 self._valid_structure["columns"], cols))
 
     def verify_schema(self):
-        conn = sqlite3.Connection(self.db)
+        """Match the selected DB to the class's valid schema to verify
+        compatibility."""
+        conn = _sqlite3.Connection(self.db)
         cur = conn.cursor()
         try:
-            self._correct_tables_and_rows(cur)
-        except AssertionError as err:
-            print(str(err))
-            return False
+            self._correct_tables_and_titles(cur)
         finally:
             conn.close()
-        return True
 
     def find_db(self):
-        cookies_path = os.path.expanduser(self._db_paths[sys.platform])
+        """Look at the default profile path based on system platform to
+        find the browser's cookie database."""
+        cookies_path = _os.path.expanduser(self._db_paths[_sys.platform])
 
         cookies_path = self._find_db_extra(cookies_path)
 
-        if not os.path.exists(cookies_path):
-            raise OSError("Cookie does not exist at {}".format(cookies_path))
+        if not _os.path.exists(cookies_path):
+            raise BrowserException("Cookie does not exist at "
+                                   "{}".format(cookies_path))
         return cookies_path
 
     @staticmethod
     def _find_db_extra(expanded_path):
+        """Some browser's profiles require path manipulation"""
         return expanded_path
 
     @staticmethod
-    def _current_time(epoch=datetime.datetime(1970, 1, 1), length=16):
-        delta_from_epic = (datetime.datetime.utcnow() - epoch)
-        return str(delta_from_epic.total_seconds()
-                   ).replace(".", "")[:length].zfill(length)
+    def _current_time(epoch=_dt.datetime(1970, 1, 1), length=16):
+        """Returns a string of the current time based on epoc date at a set
+         length of integers."""
+        delta_from_epic = (_dt.datetime.utcnow() - epoch)
+        return int(str(_to_secs(delta_from_epic)
+                       ).replace(".", "")[:length].ljust(length, "0"))
 
     @staticmethod
-    def _expire_time(epoch=datetime.datetime(1970, 1, 1), length=10, days=365):
-        delta_from_epic = (datetime.datetime.utcnow() - epoch
-                           + datetime.timedelta(days=days))
-        return str(delta_from_epic.total_seconds()
-                   ).replace(".", "")[:length].zfill(length)
+    def _expire_time(epoch=_dt.datetime(1970, 1, 1), length=10, days=365,
+                     hours=0):
+        """Returns a string of time based on epoch date at a set
+         length of integers with an additional timedetla of specified days
+        and/or hours."""
+        delta_from_epic = (_dt.datetime.utcnow() - epoch +
+                           _dt.timedelta(days=days, hours=hours))
+        return int(str(_to_secs(delta_from_epic)
+                       ).replace(".", "")[:length].ljust(length, "0"))
 
     def _insert_command(self, cursor, host, name, value, path,
                         expires_at, secure, http_only, **extra):
+        """Child class must override"""
         raise NotImplementedError()
 
     def _delete_command(self, cursor, host, name):
+        """Child class must override"""
         raise NotImplementedError()
 
     def add_cookie(self, host, name, value, path="/", expires_at=None, secure=0,
                    http_only=0, **extra):
-        conn = sqlite3.Connection(self.db)
+        """Abstracted function to add a cookie to the database."""
+        conn = _sqlite3.Connection(self.db)
         cur = conn.cursor()
         try:
             self._insert_command(cur, host, name, value, path, expires_at,
@@ -97,12 +147,13 @@ class CookieManager(object):
             conn.close()
 
     def delete_cookie(self, host, name):
-        conn = sqlite3.Connection(self.db)
+        """Abstracted function to remove a cookie from the database."""
+        conn = _sqlite3.Connection(self.db)
         cur = conn.cursor()
         try:
             self._delete_command(cur, host, name)
         except Exception as err:
-            raise err
+            raise BrowserException(str(err))
         else:
             conn.commit()
         finally:
@@ -110,14 +161,15 @@ class CookieManager(object):
 
     def update_cookie(self, host, name, value, path="/", expires_at=None,
                       secure=0, http_only=0, ignore_missing=True, **extra):
-        conn = sqlite3.Connection(self.db)
+        """Delete and re-add a cookie with different value."""
+        conn = _sqlite3.Connection(self.db)
         cur = conn.cursor()
 
         try:
             self._delete_command(cur, host, name)
         except Exception as err:
             if not ignore_missing:
-                raise err
+                raise BrowserException(str(err))
 
         try:
             self._insert_command(cur, host, name, value, path, expires_at,
@@ -130,7 +182,8 @@ class CookieManager(object):
             conn.close()
 
 
-class FirefoxCookies(CookieManager):
+class FirefoxCookiesV1(CookieManager):
+    """First iteration of Firefox Cookie manager"""
     _valid_structure = {"tables": [("moz_cookies",)],
                         "columns": ['id', 'baseDomain', 'originAttributes',
                                     'name', 'value', 'host', 'path', 'expiry',
@@ -147,17 +200,22 @@ class FirefoxCookies(CookieManager):
         "darwin": "~/Library/Application Support/Firefox/Profiles/",
         "linux": "~/.mozilla/firefox/"}
     table_name = "moz_cookies"
+    supported_versions = (47, 48)
 
     @staticmethod
     def _find_db_extra(expanded_path):
-        default = [x for x in os.listdir(expanded_path)
+        """Firefox profiles path has a folder that ends with .default that
+        must be found."""
+        default = [x for x in _os.listdir(expanded_path)
                    if x.endswith(".default")]
         if not default:
-            raise OSError("No default profile in {}".format(expanded_path))
-        return os.path.join(expanded_path, default[0], "cookies.sqlite")
+            raise BrowserException("No default profile in "
+                                   "{}".format(expanded_path))
+        return _os.path.join(expanded_path, default[0], "cookies.sqlite")
 
     def _insert_command(self, cursor, host, name, value, path,
                         expires_at, secure, http_only, **extra):
+        """Firefox specific SQL insert command with required times"""
         now = self._current_time(length=16)
         exp = self._expire_time(length=10)
         base_domain = extra.get("base_domain", ".".join(host.split(".")
@@ -171,11 +229,13 @@ class FirefoxCookies(CookieManager):
                                       extra.get('in_browser_element', 0)))
 
     def _delete_command(self, cursor, host, name):
+        """Firefox specific SQL delete command"""
         cursor.execute("DELETE FROM moz_cookies WHERE host=? AND name=?",
                        (host, name))
 
 
-class ChromeCookies(CookieManager):
+class ChromeCookiesV1(CookieManager):
+    """First iteration of Chrome Cookie manager"""
     _valid_structure = {"tables": [("meta",), ("cookies",)],
                         "columns": ['creation_utc', 'host_key', 'name', 'value',
                                     'path', 'expires_utc', 'secure', 'httponly',
@@ -193,11 +253,13 @@ class ChromeCookies(CookieManager):
                " firstpartyonly) VALUES (?, ?, ?, ?, ?, ?, ?, ?, "
                "?, ?, ?, ?, ?, ?)")
     table_name = "cookies"
+    supported_versions = (52,)
 
     def _insert_command(self, cursor, host, name, value, path,
                         expires_at, secure, http_only, **extra):
-        now = self._current_time(epoch=datetime.datetime(1601, 1, 1), length=17)
-        exp = self._expire_time(epoch=datetime.datetime(1601, 1, 1), length=17)
+        """Chrome specific SQL insert command with required times"""
+        now = self._current_time(epoch=_dt.datetime(1601, 1, 1), length=17)
+        exp = self._expire_time(epoch=_dt.datetime(1601, 1, 1), length=17)
 
         cursor.execute(self._insert, (now, host, name, value, path, exp, secure,
                                       http_only, now,
@@ -208,5 +270,14 @@ class ChromeCookies(CookieManager):
                                       extra.get('first_party_only', 0)))
 
     def _delete_command(self, cursor, host, name):
+        """Chrome specific SQL delete command"""
         cursor.execute("DELETE FROM cookies WHERE host_key=? AND name=?",
                        (host, name))
+
+
+class FirefoxCookies(FirefoxCookiesV1):
+    """Current version of Firefox cookie management"""
+
+
+class ChromeCookies(ChromeCookiesV1):
+    """Current version of Chrome cookie management"""