+++ reusables/log.py
@@ -8,7 +8,7 @@
 Logging helper functions and common log formats.
 """
 import logging as _logging
-import sys
+import sys as _sys
 
 log_common_format = '%(asctime)s - %(name)s - %(levelname)s - %(message)s'
 log_level_first_format = '%(levelname)s - %(name)s - %(asctime)s - %(message)s'
@@ -20,7 +20,7 @@ log_easy_thread_format = '%(relativeCreated)8d %(threadName)s : %(name)-12s ' \
 log_detailed_format = '%(asctime)s : %(relativeCreated)5d %(threadName)s : ' \
                       '%(name)s %(levelname)s %(message)s'
 
-if sys.version_info < (2, 7):
+if _sys.version_info < (2, 7):
     class NullHandler(_logging.Handler):
         def emit(self, record):
             pass
@@ -28,7 +28,7 @@ if sys.version_info < (2, 7):
     _logging.NullHandler = NullHandler
 
 
-def get_stream_handler(stream=sys.stderr, level=_logging.INFO,
+def get_stream_handler(stream=_sys.stderr, level=_logging.INFO,
                        log_format=log_easy_read_format):
     """
     Returns a set up stream handler to add to a logger.
@@ -60,7 +60,7 @@ def get_file_handler(file_path="out.log", level=_logging.INFO,
     return fh
 
 
-def get_logger(module_name=__name__, level=_logging.INFO, stream=sys.stderr,
+def get_logger(module_name=__name__, level=_logging.INFO, stream=_sys.stderr,
                file_path=None, log_format=log_easy_read_format,
                suppress_warning=True, ignore_existing=False):
     """